feat(errors): move errors and create new classes
