+++ src/lib/errors/Imperial/DocumentNotFound.ts
